+++ ui-packages/packages/form-details/src/embedded/EmbeddedFormDetails.tsx
@@ -23,11 +23,13 @@ import {
   FormDetailsDriver
 } from '../api';
 import { FormDetailsChannelApiImpl } from './FormDetailsChannelApiImpl';
+import { FormInfo } from '@kogito-apps/forms-list';
 import { ContainerType } from '@kogito-tooling/envelope/dist/api';
 import { init } from '../envelope';
 export interface Props {
   targetOrigin: string;
   driver: FormDetailsDriver;
+  formData: FormInfo;
 }
 
 export const EmbeddedFormDetails = React.forwardRef<FormDetailsApi, Props>(
@@ -58,8 +60,7 @@ export const EmbeddedFormDetails = React.forwardRef<FormDetailsApi, Props>(
             envelopeServerId: envelopeServer.id
           },
           {
-            // @ts-ignore
-            formData: props.formData
+            ...props.formData
           }
         );
       },
